requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
scipy==1.12.0
matplotlib==3.8.3
seaborn==0.13.2
plotly==5.20.0
google-generativeai==0.8.3
sqlalchemy==2.0.29
fpdf2==2.7.9
openpyxl==3.1.2
python-dotenv==1.0.1
xlsxwriter==3.2.0
